servicio_bot_whatsapp/requirements.txt
# Core
Flask[async]>=2.0
gunicorn>=20.0.0

# Async HTTP
httpx>=0.24.0

# Validation
pydantic>=2.0

# Redis
redis>=4.5.0

# Message Queue
pika>=1.3.0

# Twilio
twilio>=8.0.0

# AI
google-generativeai>=0.3.0
openai-whisper

# Utils
python-dotenv